[bazel] Add clangd as a library support (#81556)
Creates a `BUILD.bazel` file for the `clangd` directory in the project
overlay.

This upstreams the patch that allows
https://github.com/carbon-language/carbon-lang/tree/trunk/language_server
to use `clangd` as a library. This was created as part of a Summer of
Code project building a prototype Carbon language server. If this is not
an appropriate architecture, I'm very open to alternative paths forward.

Thanks!